requirements/base.txt
amqp==2.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
billiard==3.6.4.0
celery==4.4.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cx-Oracle==8.3.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.12
django-allauth==0.49.0
django-celery-beat==2.2.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-model-utils==4.2.0
django-redis==5.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
flower==0.9.7
future==0.18.2
h11==0.13.0
hiredis==2.0.0
httptools==0.2.0
humanize==4.0.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==4.6.11
MarkupSafe==2.1.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
prometheus-client==0.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==6.1.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tornado==6.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.14.0
vine==1.3.0
watchgod==0.7
websockets==10.2
whitenoise==6.0.0
wrapt==1.13.3
